requirements.txt
fastapi
uvicorn
insightface
opencv-python
numpy
pillow
python-multipart
onnxruntime
